+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added task to update `update.json` with current version
 - Added getMapperName to ConfigHelper
 - Add Kiln support to Immersive Engineering
+- Added option to ignore slag output in Blast Furnace for Immersive Engineering
 
 ### Changed
 - Moved from Travis to GitHub Actions